refactor: structural decomposition with regret-based regression proof
## Decomposition
- general-use.js (977→80 lines): Split into 3 focused modules
  - general-use-parse-word.js: parseWord, parseColumn, parseTehta, scanWord, abbreviations
  - general-use-parse-tengwa.js: parseTengwa character-to-tengwa mapping
  - general-use-parse-following.js: parseTengwaAnnotations, parseFollowingAbove/Below
- classical.js (611→140 lines): Split into 2 focused modules
  - classical-parse-word.js: parseWord, parseColumn, parseTehta with reverseCurls helpers
  - classical-parse-tengwa.js: parseTengwa character mapping

## Reduce Duplication
- tengwar-annatar.js & tengwar-parmaite.js: Extracted shared
  transcribe/transcribeColumn/tehtaForTengwa into font-transcriber.js factory

## Naming Improvements
- Column.prototype.varies → added markAsVariant alias (backward-compatible)
- Renamed local curlReversals map to avoid shadowing options.reverseCurls

## Regression Proof (3 Verifications)
All verifications passed before this commit:
1. Regrets validation: 10/10 clusters GREEN
2. Raw output vs KEBENARAN 1: 31/31 outputs identical
3. Fingerprints vs KEBENARAN 2: 10/10 fingerprints match

Target: kriskowal/tengwarjs
Tool: Wolfvin/Regrets (regression-testing skill)
