+++ regrets/beleriand-encode.regret
@@ -1,0 +1,12 @@
+cluster: beleriand-encode
+version: 1
+fingerprint: 2g1h41s
+captured: 2026-06-13T16:20:46.810Z
+watches: [encode]
+entry: encode
+stack: js
+fingerprintLevel: entry
+---
+INPUT  "mordor"
+OUTPUT "vala;anna;romen;ando;anna;romen"
+HASH   2g1h41s
