+++ regrets/general-use-make-options.regret
@@ -1,0 +1,12 @@
+cluster: general-use-make-options
+version: 1
+fingerprint: 4rbm6l6
+captured: 2026-06-13T16:20:46.805Z
+watches: [makeOptions]
+entry: makeOptions
+stack: js
+fingerprintLevel: entry
+---
+INPUT  null
+OUTPUT {"font":{"tengwar":{"0":"ð","1":"ñ","2":"ò","3":"ó","4":"ô","5":"õ","6":"ö","7":"÷","8":"ø","9":"ù","10":"ú","11":"û","tinco":"1","parma":"q","calma":"a","quesse":"z","ando":"2","umbar":"w","anga":"s","ungwe":"x","thule":"3","formen":"e","harma":"d","hwesta":"c","anto":"4","ampa":"r","anca":"f","unque":"v","numen":"5","malta":"t","noldo":"g","nwalme":"b","ore":"6","vala":"y","anna":"h","wilya":"n","romen":"7","arda":"u","lambe":"j","alda":"m","silme":"8","silme-nuquerna":"i","esse":"k","esse-nuquerna":",","hyarmen":"9","hwesta-sindarinwa":"o","yanta":"l","ure":".","halla":"½","short-carrier":"`","long-carrier":"~","round-carrier":"]","thuletinco":"!","formenparma":"Q","harmacalma":"A","hwestaquesse":"Z","antoando":"@","ampaumbar":"W","ancaanga":"S","unqueungwe":"X","colon":"ˆ","apostrophe":"²","hyphen":"¬","semi-colon":"Â","comma":"=","full-stop":"-","exclamation-point":"Á","question-mark":"À","open-bracket":"›","close-bracket":"›","flourish-left":"Ğ","flourish-right":"ğ","open-paren":"Œ","close-paren":"œ"},"tehtar":{"a":"#EDC","e":"$RFV","i":"%TGB","o":"^YHN","u":["&","U","J","M","Ā","ā","Ă","ă"],"ó":["Ą","ą","Ć","ć"],"ú":["Ĉ","ĉ","Ċ","ċ"],"í":["Ô","Õ","Ö","×"],"w":"èéêë","y-english":"ØÙÚÛ","y-sindarin":"ÔÕÖ×","y-quenya":"ÌÍÎÏ´","o-below":["ä","å","æ","ç","|"],"i-below":["È","É","Ê","Ë","L"],"s":{"special":true,"calma":"|","quesse":"|","short-carrier":"}"},"s-final":{"special":true,"tinco":"+","ando":"+","numen":"+","malta":null,"lambe":"_"},"s-inverse":{"special":true,"tinco":"¡"},"s-extended":{"special":true,"tinco":"Ç"},"s-flourish":{"special":true,"tinco":"£","lambe":"¥"},"tilde-above":"Pp","tilde-below":[":",";","°"],"tilde-high-above":")0","tilde-far-below":"?/","bar-above":"{[","bar-below":["\"","'","ç"],"bar-high-above":"ìî","bar-far-below":"íï"},"positions":{"tinco":{"o":3,"w":3,"others":2},"parma":{"o":3,"w":3,"others":2},"calma":{"o":3,"w":3,"u":3,"o-below":1,"others":2},"quesse":{"o":3,"w":3,"o-below":1,"others":2},"ando":{"wide":true,"e":1,"o":2,"ó":1,"ú":1,"others":0},"umbar":{"wide":true,"e":1,"o":2,"ó":1,"ú":1,"others":0},"anga":{"wide":true,"e":1,"ó":1,"ú":1,"others":0},"ungwe":{"wide":true,"e":1,"o":1,"ó":1,"ú":1,"others":0},"thule":{"others":3},"formen":3,"harma":{"e":0,"o":3,"u":7,"ó":2,"ú":2,"w":0,"others":1},"hwesta":{"e":0,"o":3,"u":7,"w":0,"others":1},"anto":{"wide":true,"ó":1,"ú":1,"others":0},"ampa":{"wide":true,"ó":1,"ú":1,"others":0},"anca":{"wide":true,"u":7,"ó":1,"ú":1,"others":0},"unque":{"wide":true,"u":7,"others":0},"numen":{"wide":true,"ó":1,"ú":1,"others":0},"malta":{"wide":true,"ó":1,"ú":1,"others":0},"noldo":{"wide":true,"ó":1,"ú":1,"others":0},"nwalme":{"wide":true,"ó":1,"ú":1,"others":0},"ore":{"e":3,"o":3,"u":3,"ó":3,"ú":3,"others":1},"vala":{"e":3,"o":3,"u":3,"ó":3,"ú":3,"others":1},"anna":{"e":3,"o":3,"u":3,"ó":2,"ú":2,"others":1},"wilya":{"e":3,"o":3,"u":3,"ó":3,"ú":3,"others":1},"romen":{"e":3,"o":3,"u":3,"ó":2,"ú":2,"y-quenya":3,"o-below":null,"i-below":3,"others":1},"arda":{"a":1,"e":3,"i":1,"o":3,"u":3,"í":1,"ó":2,"ú":2,"y-quenya":3,"o-below":null,"i-below":3,"others":0},"lambe":{"wide":true,"e":1,"y-quenya":4,"ó":1,"ú":1,"o-below":null,"i-below":4,"others":0},"alda":{"wide":true,"o-below":null,"others":1},"silme":{"y-quenya":3,"o-below":2,"i-below":2,"others":null},"silme-nuquerna":{"e":3,"o":3,"u":3,"ó":3,"ú":3,"y-quenya":null,"o-below":null,"i-below":null,"others":1},"esse":{"y-quenya":null,"others":null},"esse-nuquerna":{"e":3,"o":3,"u":3,"ó":3,"ú":3,"others":1},"hyarmen":3,"hwesta-sindarinwa":{"o":2,"u":2,"ó":1,"ú":2,"others":0},"yanta":{"e":3,"o":3,"u":3,"ó":2,"ú":2,"others":1},"ure":{"e":3,"o":3,"u":3,"ó":3,"ú":3,"others":1},"halla":{"i-below":3,"others":null},"short-carrier":3,"long-carrier":{"y-quenya":null,"o-below":null,"i-below":null,"others":3},"round-carrier":3,"thuletinco":3,"formenparma":3,"harmacalma":{"o":3,"u":7,"ó":2,"ú":2,"others":1},"hwestaquesse":{"o":0,"u":7,"others":1},"antoando":{"wide":true,"e":1,"o":2,"ó":1,"ú":1,"others":0},"ampaumbar":{"wide":true,"e":1,"o":2,"ó":1,"ú":1,"others":0},"ancaanga":{"wide":true,"e":1,"ó":1,"ú":1,"others":0},"unqueungwe":{"wide":true,"e":1,"o":1,"ó":1,"ú":1,"others":0}}},"reverseCurls":false,"medialOre":false}
+HASH   4rbm6l6
