+++ regrets/general-use-transcribe.regret
@@ -1,0 +1,12 @@
+cluster: general-use-transcribe
+version: 1
+fingerprint: 5xgljhq
+captured: 2026-06-13T16:20:46.799Z
+watches: [transcribe]
+entry: transcribe
+stack: js
+fingerprintLevel: entry
+---
+INPUT  "hello"
+OUTPUT "9jR°`N"
+HASH   5xgljhq
