+++ regrets/index-encode.regret
@@ -1,0 +1,12 @@
+cluster: index-encode
+version: 1
+fingerprint: 51owtxr
+captured: 2026-06-13T16:20:46.812Z
+watches: [encode]
+entry: encode
+stack: js
+fingerprintLevel: entry
+---
+INPUT  ["hello","general-use","annatar"]
+OUTPUT "hyarmen;lambe:e,tilde-below;short-carrier:o"
+HASH   51owtxr
