+++ regrets/notation-encode.regret
@@ -1,0 +1,12 @@
+cluster: notation-encode
+version: 1
+fingerprint: 3h1403i
+captured: 2026-06-13T16:20:46.810Z
+watches: [encode]
+entry: encode
+stack: js
+fingerprintLevel: entry
+---
+INPUT  [[[[]]]]
+OUTPUT ""
+HASH   3h1403i
